added a bunch of pseudocode and changed everything over to resemble a login.
