+++ Front End/index.html
@@ -27,36 +27,51 @@
         <h1 class="">Tests</h1>
 
     </div>
+    <!-- This is our CRUD form. It holds all 4 basics for a CRUD application. This was just for testing purposes and may not even fit into 
+    our final project at all. I just made it to test out the 4 commands on the back end. I did make it resemble a create account form so that
+    we can have one ready incase we need it.-->
 
     <div class='d-flex flex-row justify-content-center flex-wrap pl-5 pr-5'>
+        <!-- This is our user creation form. It takes all the inputs a user account would normally have and creates a new user in our accounts table.
+            Later we will have password go through passport which changes the password to a long string of letters and numbers and then when 
+            the password comes back later it will change the code back into the original password in the users cookies.-->
         <form method="POST" role="form" class="w-25 align-self-center mt-4 mb-4 shadow p-2 ml-5 mr-5">
             <div class="form-group">
-                <p>Input your name and favorite color into DB</p>
-                <label for="nameC">Name</label>
+                <p>Create your account here.</p>
+                <label for="nameC">Input Your Name</label>
                 <input type="text" class="form-control" id="nameC" aria-describedby="nameHelp"
                     placeholder="Enter Your Name">
             </div>
             <div class="form-group">
-                <label for="favoriteC">Input Your Favorite Color</label>
-                <input type="text" class="form-control" id="favoriteC" placeholder="Fave Color">
+                    <label for="emailC">Input Your Email</label>
+                    <input type="email" class="form-control" id="emailC" placeholder="Email">
+                </div>
+            <div class="form-group">
+                <label for="passwordC">Input Your Password</label>
+                <input type="password" class="form-control" id="passwordC" placeholder="Pass">
             </div>
-            <a id="Create" href="/" class="btn btn-primary w-100">Enter Into DB</a>
+            <a id="Create" href="/" class="btn btn-primary w-100">Create Account</a>
         </form>
 
+        <!--This is a form where you can update your password. It takes a users name to tell the database which row to update
+             And then takes the password change that the user wants to make.
+             Later we will have it check against the original password the user provided before allowing them to change it.-->
         <form class="w-25 align-self-center mt-4 mb-4 shadow p-2 ml-5 mr-5">
-            <p>Change your favorite color to your favorite food</p>
+            <p>Update your password here.</p>
             <div class="form-group">
                 <label for="nameU">Name</label>
                 <input type="text" class="form-control" id="nameU" aria-describedby="nameHelp"
                     placeholder="Enter Name">
             </div>
             <div class="form-group">
-                <label for="favoriteU">Favorite Food</label>
-                <input type="text" class="form-control" id="favoriteU" placeholder="Fave Food">
+                <label for="favoriteU">New Password</label>
+                <input type="text" class="form-control" id="passwordU" placeholder="New Pass">
             </div>
             <a id="update" href="./" type="submit" class="btn btn-primary w-100">Update</a>
         </form>
 
+        <!--This is the delete form. It takes a users name and deletes it from the database.
+             When we actually start making the project i will actually remove this feature as its bad practice to delete from your DB-->
         <form class="w-25 align-self-center mt-4 mb-4 shadow p-2 ml-5 mr-5">
             <h3>Delete yourself from DB. Warning irreversable.</h3>
             <div class="form-group">
@@ -68,14 +83,18 @@
         </form>
 
     </div>
+
+    <!--This Table was created with bootstrap. It is where the information from the database is put onto the page with a get command.-->
     <table class="table table-striped table-dark">
         <thead>
             <tr>
                 <th scope="col">ID</th>
                 <th scope="col">Name</th>
-                <th scope="col">Favorite Thing</th>
+                <th scope="col">Email</th>
+                <th scope="col">Password</th>
             </tr>
         </thead>
+        <!-- this is where the rows of information are being appended hooking into the Get id. It will add rows between the tbody tag.-->
         <tbody id="Get">
         </tbody>
     </table>
