+++ README.md
@@ -1,1 +1,4 @@
 # FirstProject
+
+Please do not edit gitIgnore.
+Please read through the pseudocode and try to keep any code you added psuedocoded as well.
